The solver
This solves the vat damage problem in both a static (must choose all shots at once) and dynamic (one shot at a time) scenario.

I also have a main commandline tool I'm making to pipe information in, but scope in julia seems to be weird now, so I'm trying to figure it out.